docs/index.rst: Fix URL for VICE
The viceteam.org domain was lost to a domain squatter in
2014[[1]], and is now the site of a law firm. The curent site
is vice-emu.sourceforge.net.

Also, we deep link directly into the VICE monitor documentation,
since that's likely what people will want to see when following
this link. It's easy from there to get back to the top page (just
delete `vice_12.html` in the URL bar) if that's what the user wants.

[1]: https://web.archive.org/web/20140601000000*/viceteam.org